Move i18n setup into its own file and fix test
